Add portal auth service
