+++ src/backend/apis/connection/src/oauth/utils.ts
@@ -1,22 +1,6 @@
 import { ServiceError, badRequestError, unauthorizedError } from '@lowerdeck/error';
 import { Context } from 'hono';
 
-export let urlsMatch = (url1: string, url2: string) => {
-  try {
-    let u1 = new URL(url1);
-    let u2 = new URL(url2);
-
-    return (
-      u1.protocol == u2.protocol &&
-      u1.hostname == u2.hostname &&
-      u1.port == u2.port &&
-      u1.pathname == u2.pathname
-    );
-  } catch {
-    return false;
-  }
-};
-
 export let getString = (value: unknown) => {
   if (typeof value == 'string' && value.trim().length > 0) return value;
   return undefined;
@@ -100,36 +84,4 @@ export let getClientCredentials = (c: Context, body: Record<string, string>) => 
     clientId: fromBasic?.clientId ?? bodyClientId,
     clientSecret: fromBasic?.clientSecret ?? bodyClientSecret
   };
-};
-
-export let validateUrlString = (value: string, field: string) => {
-  try {
-    new URL(value);
-  } catch {
-    throw new ServiceError(
-      badRequestError({
-        message: `${field} must be a valid URL`,
-        oauth: {
-          error: 'invalid_request',
-          errorMessage: `${field} must be a valid URL`
-        }
-      })
-    );
-  }
-};
-
-export let validateRedirectUri = (redirectUri: string, allowedRedirectUris: string[]) => {
-  validateUrlString(redirectUri, 'redirect_uri');
-
-  if (!allowedRedirectUris.some(allowedUri => urlsMatch(allowedUri, redirectUri))) {
-    throw new ServiceError(
-      badRequestError({
-        message: 'Invalid redirect URI',
-        oauth: {
-          error: 'invalid_request',
-          errorMessage: 'Invalid redirect URI'
-        }
-      })
-    );
-  }
 };
