+++ src/backend/apis/connection/src/portal.ts
@@ -1,30 +1,16 @@
-import { badRequestError, ServiceError } from '@lowerdeck/error';
 import { useValidatedBody } from '@lowerdeck/hono';
-import { generateCustomId } from '@lowerdeck/id';
 import { v } from '@lowerdeck/validation';
 import { getConfig } from '@metorial/config';
-import { db, ID } from '@metorial/db';
 import { AuthInfo } from '@metorial/module-access';
-import { Authenticator } from '@metorial/rest';
-import { addDays } from 'date-fns';
-import { getMagicMcpTokenSecretFromRequest, handleMagicMcpRequest } from './magic';
 import {
-  exchangeAuthorizationCodeToken,
-  exchangeRefreshToken,
-  getPortalAuthClient,
+  consumerOAuthService,
   portalAccessTokenTtlSeconds,
-  portalRefreshTokenTtlSeconds,
-  resolvePortalRoute,
-  validateClientSecret
-} from './oauth/client';
+  portalRefreshTokenTtlSeconds
+} from '@metorial/module-portal';
+import { Authenticator } from '@metorial/rest';
+import { getMagicMcpTokenSecretFromRequest, handleMagicMcpRequest } from './magic';
 import { createOAuthHono } from './oauth/hono';
-import {
-  getClientCredentials,
-  getString,
-  parseOAuthBody,
-  validateRedirectUri,
-  validateUrlString
-} from './oauth/utils';
+import { getClientCredentials, getString, parseOAuthBody } from './oauth/utils';
 
 export let createPortalHandler = (d: {
   authenticate: Authenticator<AuthInfo>;
@@ -32,6 +18,38 @@ export let createPortalHandler = (d: {
   metadataServer: ReturnType<typeof createOAuthHono>;
   connectPortalServer: ReturnType<typeof createOAuthHono>;
 } => {
+  let buildPortalOAuthClientRegistration = (d: {
+    registration: {
+      id: string;
+      clientId: string;
+      clientSecret: string | null;
+      redirectUris: string[];
+      name: string;
+      tokenEndpointAuthMethod: 'client_secret_basic' | 'client_secret_post' | 'none';
+      createdAt: Date;
+      expiresAt: Date;
+    };
+    base: string;
+  }) => ({
+    client_id: d.registration.clientId,
+    client_secret: d.registration.clientSecret ?? undefined,
+    redirect_uris: d.registration.redirectUris,
+    client_name: d.registration.name,
+    grant_types: ['authorization_code', 'refresh_token'],
+    response_types: ['code'],
+    token_endpoint_auth_method: d.registration.tokenEndpointAuthMethod,
+    registration_client_uri: `${d.base}/oauth/register/${d.registration.id}`,
+    client_id_issued_at: Math.floor(d.registration.createdAt.getTime() / 1000),
+    client_secret_expires_at:
+      d.registration.clientSecret == null
+        ? 0
+        : Math.floor(d.registration.expiresAt.getTime() / 1000),
+    client_secret_issued_at:
+      d.registration.clientSecret == null
+        ? undefined
+        : Math.floor(d.registration.createdAt.getTime() / 1000)
+  });
+
   let getProtectedResource = (base: string) => ({
     resource: `${base}`,
     authorization_servers: [base],
@@ -68,7 +86,7 @@ export let createPortalHandler = (d: {
         ? rawMagicMcpTargetId
         : undefined;
 
-    return await resolvePortalRoute({
+    return await consumerOAuthService.resolvePortalRoute({
       portalId,
       magicMcpTargetId
     });
@@ -159,52 +177,17 @@ export let createPortalHandler = (d: {
         })
       );
 
-      for (let redirectUri of input.redirect_uris) {
-        validateUrlString(redirectUri, 'redirect_uri');
-      }
-
-      let tokenEndpointAuthMethod = input.token_endpoint_auth_method ?? 'client_secret_basic';
-      let clientSecret =
-        tokenEndpointAuthMethod == 'none'
-          ? null
-          : await ID.generateId('portalAuthClientSecret');
-
-      let registration = await db.portalAuthClient.create({
-        data: {
-          id: await ID.generateId('portalAuthClient'),
-          portalOid: portal.oid,
-          magicMcpServerOid:
-            magicMcpTarget?.type === 'server' ? magicMcpTarget.target.oid : null,
-          magicMcpEndpointOid:
-            magicMcpTarget?.type === 'endpoint' ? magicMcpTarget.target.oid : null,
-          name: input.client_name,
+      let registration = await consumerOAuthService.registerPortalOAuthClient({
+        portal,
+        magicMcpTarget,
+        input: {
+          clientName: input.client_name,
           redirectUris: input.redirect_uris,
-          clientId: await ID.generateId('portalAuthClientId'),
-          clientSecret,
-          tokenEndpointAuthMethod,
-          expiresAt: addDays(new Date(), 30)
+          tokenEndpointAuthMethod: input.token_endpoint_auth_method
         }
       });
 
-      return c.json({
-        client_id: registration.clientId,
-        client_secret: registration.clientSecret ?? undefined,
-        redirect_uris: registration.redirectUris,
-        client_name: registration.name,
-        grant_types: ['authorization_code', 'refresh_token'],
-        response_types: ['code'],
-        token_endpoint_auth_method: registration.tokenEndpointAuthMethod,
-        registration_client_uri: `${base}/oauth/register/${registration.id}`,
-        client_id_issued_at: Math.floor(registration.createdAt.getTime() / 1000),
-        client_secret_expires_at:
-          registration.clientSecret == null
-            ? 0
-            : Math.floor(registration.expiresAt.getTime() / 1000),
-        client_secret_issued_at:
-          registration.clientSecret == null
-            ? undefined
-            : Math.floor(registration.createdAt.getTime() / 1000)
-      });
+      return c.json(buildPortalOAuthClientRegistration({ registration, base }));
     });
   }
 
@@ -213,128 +196,21 @@ export let createPortalHandler = (d: {
     ':portalId/oauth/authorize'
   ]) {
     connectPortalServer = connectPortalServer.get(path, async c => {
-      let { portal, magicMcpTarget, portalUrl } = await resolveRoute(c);
-
-      let responseType = getString(c.req.query('response_type'));
-      let clientId = getString(c.req.query('client_id'));
-      let redirectUri = getString(c.req.query('redirect_uri'));
-      let codeChallenge = getString(c.req.query('code_challenge'));
-      let codeChallengeMethod = getString(c.req.query('code_challenge_method'));
-      let state = getString(c.req.query('state'));
-
-      if (!responseType) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'response_type is required',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'response_type is required'
-            }
-          })
-        );
-      }
-
-      if (responseType != 'code') {
-        throw new ServiceError(
-          badRequestError({
-            message: 'Only response_type=code is supported',
-            oauth: {
-              error: 'unsupported_response_type',
-              errorMessage: 'Only response_type=code is supported'
-            }
-          })
-        );
-      }
-
-      if (!clientId) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'client_id is required',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'client_id is required'
-            }
-          })
-        );
-      }
-
-      if (!redirectUri) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'redirect_uri is required',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'redirect_uri is required'
-            }
-          })
-        );
-      }
-
-      if (codeChallengeMethod && !['S256', 's256', 'none'].includes(codeChallengeMethod)) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'Only S256 PKCE challenges are supported',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'Only S256 PKCE challenges are supported'
-            }
-          })
-        );
-      }
-
-      let normalizedCodeChallengeMethod =
-        codeChallengeMethod == 'S256' || codeChallengeMethod == 's256'
-          ? ('s256' as const)
-          : codeChallenge
-            ? ('s256' as const)
-            : ('none' as const);
-
-      if (normalizedCodeChallengeMethod == 's256' && !codeChallenge) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'code_challenge is required when using PKCE',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'code_challenge is required when using PKCE'
-            }
-          })
-        );
-      }
-
-      let client = await getPortalAuthClient({
-        clientId,
-        portalOid: portal.oid,
-        magicMcpServerOid:
-          magicMcpTarget?.type === 'server' ? magicMcpTarget.target.oid : undefined,
-        magicMcpEndpointOid:
-          magicMcpTarget?.type === 'endpoint' ? magicMcpTarget.target.oid : undefined
-      });
-      validateRedirectUri(redirectUri, client.redirectUris);
-
-      let attempt = await db.portalAuthAttempt.create({
-        data: {
-          id: await ID.generateId('portalAuthAttempt'),
-          portalAuthClientOid: client.oid,
-          status: 'pending',
-          redirectUri,
-          state,
-          authorizationCode: generateCustomId('prtl_oatc_', 35),
-          codeChallengeMethod: normalizedCodeChallengeMethod,
-          codeChallenge,
-          expiresAt: addDays(new Date(), 7)
+      let { portal, magicMcpTarget } = await resolveRoute(c);
+      let authorization = await consumerOAuthService.createPortalOAuthAuthorization({
+        portal,
+        magicMcpTarget,
+        input: {
+          responseType: getString(c.req.query('response_type')),
+          clientId: getString(c.req.query('client_id')),
+          redirectUri: getString(c.req.query('redirect_uri')),
+          codeChallenge: getString(c.req.query('code_challenge')),
+          codeChallengeMethod: getString(c.req.query('code_challenge_method')),
+          state: getString(c.req.query('state'))
         }
       });
 
-      let redirectUrl = new URL(portalUrl);
-      let basePath = redirectUrl.pathname.replace(/\/+$/, '');
-      redirectUrl.pathname = `${basePath}/oauth/authorize/${attempt.id}`.replace(
-        /\/{2,}/g,
-        '/'
-      );
-      redirectUrl.search = '';
-      redirectUrl.hash = '';
-
-      return c.redirect(redirectUrl.toString(), 302);
+      return c.redirect(authorization.redirectUrl, 302);
     });
   }
 
@@ -344,113 +220,19 @@ export let createPortalHandler = (d: {
 
       let body = await parseOAuthBody(c);
       let credentials = getClientCredentials(c, body);
-      let grantType = getString(body.grant_type);
-      let code = getString(body.code);
-      let redirectUri = getString(body.redirect_uri);
-      let codeVerifier = getString(body.code_verifier);
-      let refreshToken = getString(body.refresh_token);
-
-      if (!grantType) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'grant_type is required',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'grant_type is required'
-            }
-          })
-        );
-      }
-
-      if (!credentials.clientId) {
-        throw new ServiceError(
-          badRequestError({
-            message: 'client_id is required',
-            oauth: {
-              error: 'invalid_request',
-              errorMessage: 'client_id is required'
-            }
-          })
-        );
-      }
-
-      let client = await getPortalAuthClient({
-        clientId: credentials.clientId,
-        portalOid: portal.oid,
-        magicMcpServerOid:
-          magicMcpTarget?.type === 'server' ? magicMcpTarget.target.oid : undefined,
-        magicMcpEndpointOid:
-          magicMcpTarget?.type === 'endpoint' ? magicMcpTarget.target.oid : undefined
+      let tokens = await consumerOAuthService.exchangePortalOAuthToken({
+        portal,
+        magicMcpTarget,
+        input: {
+          clientId: credentials.clientId,
+          clientSecret: credentials.clientSecret,
+          grantType: getString(body.grant_type),
+          code: getString(body.code),
+          redirectUri: getString(body.redirect_uri),
+          codeVerifier: getString(body.code_verifier),
+          refreshToken: getString(body.refresh_token)
+        }
       });
-      validateClientSecret({
-        client,
-        clientSecret: credentials.clientSecret
-      });
-
-      let tokens: { accessToken: string; refreshToken: string };
-      if (grantType == 'authorization_code') {
-        if (!code) {
-          throw new ServiceError(
-            badRequestError({
-              message: 'code is required for the authorization_code grant',
-              oauth: {
-                error: 'invalid_request',
-                errorMessage: 'code is required for the authorization_code grant'
-              }
-            })
-          );
-        }
-
-        if (!redirectUri) {
-          throw new ServiceError(
-            badRequestError({
-              message: 'redirect_uri is required for the authorization_code grant',
-              oauth: {
-                error: 'invalid_request',
-                errorMessage: 'redirect_uri is required for the authorization_code grant'
-              }
-            })
-          );
-        }
-
-        tokens = await exchangeAuthorizationCodeToken({
-          portal,
-          magicMcpTarget,
-          client,
-          code,
-          redirectUri,
-          codeVerifier
-        });
-      } else if (grantType == 'refresh_token') {
-        if (!refreshToken) {
-          throw new ServiceError(
-            badRequestError({
-              message: 'refresh_token is required for the refresh_token grant',
-              oauth: {
-                error: 'invalid_request',
-                errorMessage: 'refresh_token is required for the refresh_token grant'
-              }
-            })
-          );
-        }
-
-        tokens = await exchangeRefreshToken({
-          portal,
-          magicMcpTarget,
-          client,
-          refreshToken
-        });
-      } else {
-        throw new ServiceError(
-          badRequestError({
-            message: `Unsupported grant type: ${grantType}`,
-            oauth: {
-              error: 'unsupported_grant_type',
-              errorMessage: `Unsupported grant type: ${grantType}`
-            }
-          })
-        );
-      }
 
       return c.json({
         access_token: tokens.accessToken,
@@ -469,16 +251,10 @@ export let createPortalHandler = (d: {
     connectPortalServer = connectPortalServer.get(path, async c => {
       let registrationId = c.req.param('registrationId');
       let { portal, magicMcpTarget, base } = await resolveRoute(c);
-
-      let registration = await db.portalAuthClient.findFirst({
-        where: {
-          id: registrationId,
-          portalOid: portal.oid,
-          magicMcpServerOid:
-            magicMcpTarget?.type === 'server' ? magicMcpTarget.target.oid : null,
-          magicMcpEndpointOid:
-            magicMcpTarget?.type === 'endpoint' ? magicMcpTarget.target.oid : null
-        }
+      let registration = await consumerOAuthService.getPortalOAuthRegistration({
+        portal,
+        magicMcpTarget,
+        registrationId
       });
 
       if (!registration) {
@@ -491,25 +267,7 @@ export let createPortalHandler = (d: {
         );
       }
 
-      return c.json({
-        client_id: registration.clientId,
-        client_secret: registration.clientSecret ?? undefined,
-        redirect_uris: registration.redirectUris,
-        client_name: registration.name,
-        grant_types: ['authorization_code', 'refresh_token'],
-        response_types: ['code'],
-        token_endpoint_auth_method: registration.tokenEndpointAuthMethod,
-        registration_client_uri: `${base}/oauth/register/${registration.id}`,
-        client_id_issued_at: Math.floor(registration.createdAt.getTime() / 1000),
-        client_secret_expires_at:
-          registration.clientSecret == null
-            ? 0
-            : Math.floor(registration.expiresAt.getTime() / 1000),
-        client_secret_issued_at:
-          registration.clientSecret == null
-            ? undefined
-            : Math.floor(registration.createdAt.getTime() / 1000)
-      });
+      return c.json(buildPortalOAuthClientRegistration({ registration, base }));
     });
   }
