+++ src/backend/modules/portal/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './portal';
+export * from './consumerOAuth';
